strlen updated

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: despanad <despanad@student.42barcelona.co  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/11 22:44:40 by despanad          #+#    #+#             */
-/*   Updated: 2024/06/11 22:49:34 by despanad         ###   ########.fr       */
+/*   Updated: 2024/06/21 19:37:55 by despanad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
 	int	index;
